move aws util to own class
